refactor(dataverse): reoganize vc related code
