+++ contracts/okp4-dataverse/src/credential/mod.rs
@@ -1,0 +1,3 @@
+mod proof;
+mod rdf_markers;
+mod vc;
