--- contracts/okp4-dataverse/src/did/mod.rs
@@ -1,3 +1,0 @@
-mod consts;
-mod crypto;
-mod vc;
